requirements.txt
Flask==3.1.1
Pillow==11.2.1
Requests==2.32.4
